3year/5сем/РСЧИР/Пр11-12/pr11/rschir11/requirements.txt
asgiref==3.7.2
contourpy==1.2.0
cycler==0.12.1
Django==4.2.7
djangorestframework==3.14.0
fonttools==4.45.0
importlib-resources==6.1.1
kiwisolver==1.4.5
matplotlib==3.8.2
numpy==1.26.2
packaging==23.2
Pillow==10.1.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
zipp==3.17.0
